% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pemaps.R
\docType{data}
\name{departamental}
\alias{departamental}
\title{Base de datos para graficar a nivel departamental el mapa de Peru.}
\format{Base de datos para graficar a nivel departamento el mapa de Peru.}
\usage{
data(departamental)
}
\description{
Base de datos para graficar a nivel departamental el mapa de Peru.
}
\keyword{datasets}
