% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pemaps.R
\docType{data}
\name{edificaciones}
\alias{edificaciones}
\title{Base de datos para graficar a nivel edificaciones el mapa de Peru.}
\format{Base de datos para graficar a nivel edificaciones el mapa de Peru.}
\usage{
data(edificaciones)
}
\description{
Base de datos para graficar a nivel edificaciones el mapa de Peru.
}
\keyword{datasets}
