% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pemaps.R
\docType{data}
\name{provincial}
\alias{provincial}
\title{Base de datos para graficar a nivel provincial el mapa de Peru.}
\format{Base de datos para graficar a nivel provincial el mapa de Peru.}
\usage{
data(provincial)
}
\description{
Base de datos para graficar a nivel provincial el mapa de Peru.
}
\keyword{datasets}
